hey
hey

+++ Expression.h
@@ -5,8 +5,8 @@ using namespace std;
  * Expression Interface
  */
 class Expression {
- public:
-  virtual double calculate() = 0;
-  virtual ~Expression(){}
+public:
+    virtual float calculate() = 0;
+    virtual ~Expression(){}
 };
 #endif //PROJECT_EXPRESSION_H
